requirements.txt
python-telegram-bot==13.7
requests==2.32.3
python-dotenv==1.1.0
watchdog==6.0.0
coinmarketcap==5.0.3
requests_cache==1.2.1
attrs==25.3.0
platformdirs==4.3.7
cattrs==24.1.3
url-normalize==2.2.0
pytz>=2018.6
cachetools==4.2.2
tornado>=6.1
APScheduler==3.6.3
six>=1.4.0
tzlocal>=1.2
python-twitter==3.5
pandas==2.1.1
plotly==5.19.0
beautifulsoup4==4.12.2
